Remove MatrixGroup protocol and refactor SyncGroup usage
Deleted MatrixGroup.swift and moved its logic directly into SyncGroup. Updated SyncCoordinator to use SyncGroup instead of MatrixGroup. This simplifies the codebase by consolidating matrix-related functionality into SyncGroup.

Rename MatrixGroupTests to SyncGroupTests and update usage

Renamed MatrixGroupTests.swift to SyncGroupTests.swift and updated references from TestGroup (MatrixGroup) to SyncGroup in test files. This aligns test naming and usage with the SyncGroup type, reflecting recent changes in the codebase.
